backend/requirements.txt
Django==4.0.7
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-filter==22.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
reportlab==3.6.11
